move the leader election integration test to a dedicated package
