+++ test/integration/util/leaderelection/testdata/apiserver-compose.yaml
@@ -1,0 +1,46 @@
+version: '2.3'
+services:
+  etcd:
+    image: "datadog/docker-library:etcd_3_2_6"
+    network_mode: ${network_mode}
+    environment:
+      - ETCDCTL_API=3
+    healthcheck:
+      test: ["CMD", "etcdctl", "--command-timeout=2s", "--dial-timeout=2s", "--endpoints", "http://127.0.0.1:2379", "endpoint", "health"]
+      interval: 5s
+      timeout: 5s
+      retries: 30
+
+  apiserver:
+    image: "datadog/docker-library:hyperkube_1_8_3"
+    command: "/hyperkube
+        apiserver
+        --apiserver-count=1
+        --insecure-bind-address=0.0.0.0
+        --insecure-port=8080
+        --service-cluster-ip-range=192.168.1.1/24
+        --admission-control=NamespaceLifecycle,LimitRanger,DefaultStorageClass,ResourceQuota
+        --authorization-mode=AlwaysAllow
+        --etcd-servers=http://127.0.0.1:2379"
+    network_mode: ${network_mode}
+    depends_on:
+      etcd:
+        condition: service_healthy
+    healthcheck:
+      test: ["CMD", "/hyperkube", "kubectl", "get", "cs"]
+      interval: 5s
+      timeout: 5s
+      retries: 30
+
+  pause:
+    #
+    # This pause container is here to wait until the apiserver
+    # is healthy before returning.
+    #
+    image: "datadog/docker-library:pause_3_1"
+    depends_on:
+      etcd:
+        condition: service_healthy
+      apiserver:
+        condition: service_healthy
+    network_mode: none
