+++ test/integration/util/leaderelection/testdata/kubeconfig.json
@@ -1,0 +1,31 @@
+{
+   "apiVersion": "v1",
+   "kind": "Config",
+   "clusters": [
+      {
+         "cluster": {
+            "server": "http://127.0.0.1:8080"
+         },
+         "name": "kubernetes"
+      }
+   ],
+   "contexts": [
+      {
+         "context": {
+            "cluster": "kubernetes",
+            "user": "kubelet"
+         },
+         "name": "kubelet-to-kubernetes"
+      }
+   ],
+   "current-context": "kubelet-to-kubernetes",
+   "users": [
+      {
+         "name": "kubelet",
+         "user": {
+            "password": "some-password",
+            "username": "exp"
+         }
+      }
+   ]
+}
